add priority queue to reconciling work in apply_controller
apply_controllers in the memberagent currently watches all the Work resources created in the designated namespace on the Hub cluster and reconciles them in order. This can result in serious bottleneck in applying newer resources or newer updates to existing resources over time especially when there are 1000s of namespaces created a long time ago and never have any updates to them, yet reconciled one by one before the controller can get to the newer ones.

Fixes #

This PR integrates a priority queue for adding work objects to the watch list to prioritize work objects that are newer, have recent updates or have their conditions not Ready yet.

Note that the PR uses both For and Watches in the controller setup. This is needed to avoid a runtime panic when using just Watches

This can also be done using just For without Watches, but that model would not use a PriorityQueue and just plays directly with the events. With that model, we will only watch the changed/new work objects and completely ignore old objects. With the PriorityQueue, we are actually adding every single Work object to the queue just in a prioritized order.

Signed-off-by: Sudheer Vinukonda <sudheerv@apache.org>

Signed-off-by: Sudheer Vinukonda <sudheerv@apache.org>